src/third_party/blink/web_tests/vr/getFrameData_noupdate.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="resources/vr-test-utils.js "></script>
<script src="resources/test-constants.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t, device_controller) => {
  return navigator.getVRDisplays().then( (displays) => {
    var display = displays[0];
    device_controller.setPose(VALID_POSE);
    var fd = new VRFrameData();
    t.step( () => {
      assert_false(display.getFrameData(fd));
    }, "getFrameData did not update object initially");

    function checkLater() {
      t.step( () => {
        assert_false(display.getFrameData(fd));
      }, "getFrameData did not update object after delay");
      t.done();
    }

    window.setTimeout(checkLater, 500);
  }, (err) => {
    t.step( () => {
      assert_unreached("getVRDisplays rejected");
    });
    t.done();
  });
}, [fakeDisplays["Pixel"]],
"getFrameData does not update objects until data is available after rAF call");

</script>
